refactoring to avoid duplicate code

+++ src/renderer/hw.rs
@@ -3,7 +3,7 @@ const MODE_4_PAGE_2: *mut u16 = 0x600_A000 as *mut u16;
 const REG_DISPCNT: *mut u32 = 0x0400_0000 as *mut u32;
 const DCNT_PAGE: u32 = 0x0010;
 
-pub fn draw_wide_point(x: i32, y: i32, color: u8, page: u32) {
+pub fn draw_wide_point(x: i32, y: i32, color: u32, page: u32) {
     let index = (y * 240 + x) / 2;
     let value = ((color as u16) << 8) | (color as u16);
 
@@ -26,7 +26,7 @@ pub fn fill(page: u32, color: u8) {
 
     unsafe {
         for i in 0..19200 {
-            *active_page.add(i as usize) = value;
+            *active_page.add(i) = value;
         }
     }
 }
